+++ NEWS.md
@@ -11,6 +11,8 @@
 
 - Added a `--no-progress` option.
 
+- Added a `--debug` option.
+
 ## v0.6.15 2021-01-23
 
 - Find referenced blocks by walking all bands in parallel. This significantly